requirements.txt
pandas
seaborn
streamlit
pyarrow
plotly==5.23.0
scikit-learn
idendrogram
matplotlib
scipy
pillow==9.0.0
streamlit-pdf-viewer
openai
streamlit_chat
python-dotenv
